third_party/blink/web_tests/fast/forms/select/select-assign-null-expected.txt

Check if an option is removed when it is assigned null.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tags.options.length is 2
PASS tags.options.length is 1
PASS tags.options.length is 2
PASS successfullyParsed is true

TEST COMPLETE

blink
bgsound
